% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hilmo_import_xlsx.R
\name{Hilmo_import_xlsx}
\alias{Hilmo_import_xlsx}
\title{A Hilmo import .xlsx function}
\usage{
Hilmo_import_xlsx(filename)
}
\arguments{
\item{filename}{the name of your file you wish to import}
}
\description{
This function allows you to import xlsx files in more tidy format and it creates ID column from lahtopaiva and tnro
}
\details{
Remember that this import function has different order in the lahtopvm column and if you have problems,
it should be changed depending on the data
}
\examples{
Hilmo_import_xlsx()
}
\keyword{hilmo}
\keyword{import}
\keyword{xlsx}
